adjust dummy class

+++ src/soft_cut/classification/caffe_classifier.hpp
@@ -48,7 +48,7 @@ namespace sbd {
 
 		void initializeCaffeNet(bool cpu, std::string modelFile, std::string protoFile){};
 		void predict(std::vector<cv::Mat> image, std::vector<int> labels, std::string resultLayer,
-			std::string dataLayer, std::vector<float>& predictions){};
+			std::string dataLayer, std::vector<short>& predictions){};
 	};
 
 }
